LayoutTests/fast/js/dfg-check-two-structures-expected.txt

This tests that a polymorphic structure check is actually executed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 44
PASS foo(o, o) is 42
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS foo(o, o) is 46
PASS successfullyParsed is true

TEST COMPLETE
